server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pdfplumber
python-multipart
sarvamai
